pull data hooks up into a common interface
